fix: move cn utils around
